move LSP extension into src/ and wire into default dist
extensions/lsp.ts and extensions/lsp/* were kept outside src/ for
testing isolation. Move them into src/extensions/ so they are included
in the default distribution and registered as a built-in extension.
